Add metadata generation scripts, initial result
Script that generates from the github repo needs a token to be effective (rate limiting on the api calls)
Script that generates from a local folder assumes that you have the CustomBattlers folder downloaded.

This should allow for more efficient querying of whether a set of custom sprites exist, as well as how many alts that would exist for the head/body pair. 

As is, (with no prepopulation) it won't list the pair if there is no sprite for it at all, allowing for quick querying of how many body sprites exist for each head, and relatively simple processing/querying to get the count of the head sprites for a given body.